+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Version 0.1.2.0
 *   Actually implemented the Hilbert-Huang transform
 *   Allowed for other border handling behaviors during EMD
 *   Changed default stopping conditions for sifting process
+*   Added clamped spline end conditions.
 *   Removed unsized interface
 *   Sifting will now throw runtime exception for singular splining matrices,
     instead of treating the result as a residual.  This might change in the